backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==10.1.0
numpy==1.26.4
scipy==1.11.4
